requirements.txt
langchain>=1.0.0
langgraph>=1.0.0
langchain-core>=1.0.0
langchain-openai>=0.3.0
langchain-ollama>=0.3.0
langchain-huggingface>=0.3.0
langchain-chroma>=0.2.0

gradio>=6.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
numpy>=1.26.0
rank-bm25>=0.2.2
sentence-transformers>=3.0.0
psycopg2-binary>=2.9.9
tavily-python>=0.7.0
python-dotenv>=1.0.1
loguru>=0.7.2
langchain-text-splitters>=1.0.0
